requirements.txt
langchain
ipykernel
langchain-classic
langchain-community
langchain-core
python-dotenv
streamlit
langchain-openai
pypdf
langchain-text-splitters
faiss-cpu
chromadb
langchain_chroma
langchain_groq
langchain_huggingface 
arxiv
wikipedia
ollama
langchain-ollama
langchain-tools
google-generativeai